trunk/commons-validation/src/main/java/no/schibstedsok/commons/validators/hibernate/PhoneValidator.java
/*
 * Copyright (2005-2007) Schibsted Søk AS
 */
package no.schibstedsok.commons.validators.hibernate;

import org.apache.log4j.Logger;
import org.hibernate.validator.Validator;

/**
 * {@inheritDoc}
 *
 * @author <a href="mailto:endre@sesam.no">Endre Midtgård Meckelborg</a>
 * @author <a href="mailto:ola@sesam.no">Ola Sagli</a>
 * @version <tt>$Revision: $</tt>
 */
public class PhoneValidator implements Validator<Phone> {

    /** Logger for this class. */
    private static final Logger LOG = Logger.getLogger(PhoneValidator.class);

    /** Phone type to use when validating. */
    private PhoneType phoneType = null;

    /**
     * Initialize the validator.
     *
     * @param parameters the arguments sent to the validator
     */
    public void initialize(final Phone parameters) {
        phoneType = parameters.phoneType();
    }

    /**
     * The validate method.
     *
     * @param value the value to validate
     * @return the validation result
     */
    public boolean isValid(final Object value) {
        if (value == null) {
            return true;
        }
        if (!(value instanceof String)) {
            return false;
        }

        return phoneType.getPattern().matcher((String) value).matches();
    }

}
